requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
loguru==0.7.2

# Уже есть (из текущего проекта):
curl-cffi
beautifulsoup4
openpyxl
pydantic
requests
tomli-w
tzlocal
playwright
playwright-stealth
httpx
socksio
PySocks